Fixed bans *
This update is for steam only, they are different offsets ATM and I can't test origin. FOV is locked to 10 ADS, and 50 not aiming, changed it in the client side: main.cpp line 1303 to 1319.

I made it to level 60, got to plat in ranked and not banned. Not sure after that. My fork is done, no more features unless I see something interesting. Or if someone wants to add Item ID's that work so I can get rid of item names.

This fork is feature complete at the moment. Most likely won't update offsets.

*: not sure